backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
groq>=0.10.0
langchain==0.1.0
langchain-groq>=0.0.1
google-cloud-dialogflow>=2.16.0
google-auth>=2.25.0
